requirements.txt
# Core ARI and WebSocket dependencies
ari==0.1.3
websockets==12.0
aiohttp==3.12.14

# Configuration and logging
pydantic==2.5.3
pydantic_settings==2.1.0
structlog==24.1.0
prometheus-client==0.20.0
PyYAML==6.0.1

# Utilities
tenacity==8.2.3

# Optional dependencies for audio processing
numpy>=1.21.0

# WebRTC VAD for robust speech detection
webrtcvad==2.0.10

# Email tools dependencies
resend==0.8.0
jinja2==3.1.2
dnspython==2.4.2
